new class based layout with swapParser class.
