+++ start.js
@@ -1,5 +1,5 @@
-import { Watcher } from './bots/watcher_ethers.js'
-import wallets from "./bots/wallets.js"
+import { BlockPoster } from './bots/blockPoster.js'
+import wallets from "./utils/wallets.js"
 import express from "express"
 import cors from "cors"
 import router from './routes/index.js'
@@ -16,7 +16,7 @@ app.use(router);
 app.use(cors());
 
 app.all("*", (req, res, next) => {
- next(new AppError(`The URL ${req.originalUrl} does not exists`, 404));
+ next(new AppError(`The URL ${req.originalUrl} does not exist`, 404));
 });
 app.use(errorHandler);
 const PORT = 3000;
@@ -28,15 +28,17 @@ app.listen(PORT, () => {
 const CHAT_ID_CHANNEL = -1001855095247;
 const CHAT_ID_DISCUSSION = -1001882676825;
 const CHAT_ID_CHANNEL_BETA = -1001839931719;
+const CHAT_ID_BETA_TEST = -896152552
 const testnetStatus = false
 const fullNodeIp = "192.168.0.228"
 const archiveNodeIp = "192.168.0.155"
 const httpPort = "9535"
+const chatId = CHAT_ID_BETA_TEST;
 //const wssPort = "9536" not using wss atm
 
 const fullNodeUrl = `http://${fullNodeIp}:${httpPort}`
 const archiveUrl = `http://${archiveNodeIp}:${httpPort}`
-//const archiveUrl = "http://192.168.0.155:9535";
-
-const watcher = new Watcher(CHAT_ID_CHANNEL, wallets, AlertBotKey, VolumeBotKey, testnetStatus, fullNodeUrl, archiveUrl);
 //watcher.runVolumeCheck(1)
+
+const blockPoster = new BlockPoster(chatId, wallets, AlertBotKey, fullNodeUrl, archiveUrl)
+blockPoster.start();
